requirements.txt
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
bcrypt==4.0.1
bleach==6.2.0
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.2
cryptography==45.0.7
cyclonedx-python-lib==9.1.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.116.1
filelock==3.19.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
license-expression==30.4.4
limits==5.4.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.1
nest-asyncio==1.6.0
packageurl-python==0.17.5
packaging==25.0
parso==0.8.4
passlib==1.7.4
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.10
psycopg-binary==3.2.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-serializable==2.1.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.5.2
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pywin32==310
PyYAML==6.0.2
pyzmq==27.0.0
requests==2.32.5
rich==14.1.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.3
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.2
